+++ docs/RELEASES.md
@@ -15,7 +15,7 @@ This is the list of project releases, past and current. To see tasks that are st
 - :ant:: the developer is improving stuff or fixing some sh\*t it did (increases last number)
 
 ## <a name="9.0.0.0"></a>9.0.0.0 :dragon: <sup>`5`</sup>
-- [ ] Fix version program option `-g`
+- [ ] `220719>......` Fix version program option `-g`
 - [x] `220715>220718` Add Mono themes - Dark and Light - ACCESSIBILITY
 - [x] `220715>220715` warning about schedules not running if not active
 - [x] `220715>220715` warning about contract being only with who signs, not related people